Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/navigation-timing/test_document_open-expected.txt

Description

This test validates window.performance.timing remains constant when a document is replaced using document.open.

This page should be loaded with a yellow frame below. It then replaces the document in that frame with a green document.

The test passes if all of the checks to performance.timing are correct and the frame below ends with a green background.


PASS window.performance is defined
PASS connectEnd is the same after document open.
PASS connectStart is the same after document open.
PASS domComplete is the same after document open.
PASS domContentLoadedEventEnd is the same after document open.
PASS domContentLoadedEventStart is the same after document open.
PASS domInteractive is the same after document open.
PASS domLoading is the same after document open.
PASS domainLookupEnd is the same after document open.
PASS domainLookupStart is the same after document open.
PASS fetchStart is the same after document open.
PASS loadEventEnd is the same after document open.
PASS loadEventStart is the same after document open.
PASS navigationStart is the same after document open.
PASS redirectEnd is the same after document open.
PASS redirectStart is the same after document open.
PASS requestStart is the same after document open.
PASS responseEnd is the same after document open.
PASS responseStart is the same after document open.
PASS unloadEventEnd is the same after document open.
PASS unloadEventStart is the same after document open.
